requirements.txt
streamlit
gspread
pandas
google-api-python-client
google-auth
openpyxl
fpdf
num2words